+++ mftool-src/main.c
@@ -37,7 +37,7 @@ struct var_t {
 
 static struct tf_file_header_t fseqResize(const struct tf_file_header_t header,
                                           const struct var_t *const vars) {
-    uint16_t varDataSize = 0;
+    unsigned int varDataSize = 0;
 
     for (size_t i = 0; i < arrlenu(vars); i++)
         varDataSize += sdslen(vars[i].string) + VAR_HEADER_SIZE + 1;
